Improve shader editor save functionality
In #55, it was reported that saving the page as HTML doesn't work
as expected. This commit does not directly solve the problem, but
it does make saving while the shader editor has focus more
convenient. It overrides CodeMirror's default "save" command
instead of always using "Ctrl+S". This captures the event,
prevents the default page save action, and maps to the correct
key combination if the user is on Windows/Linux/Mac. This commit
also adds a button to make it super clear you can press the button
to live update a shader.