Use a simpler `clearValue`
This is obviously a personal preference but, ... I feel like
the `{ r: ?, g: ?, b: ?, a: ? }` format is left over from before
the time when an array of numbers was allowed?

The array of numbers is arguably more useful since all kinds of things
can easily provide an array of numbers. So, it seems like it would be good
for the samples to show the more useful usage. That way devs are not
mislead into thinking they need to take some array of numbers and
convert it into the more verbose format.