Fix toneMapping reference (#455)
AFAICT, the spec requires toneMapping period. Even if the browser
doesn't support toneMapping.mode = 'extended' it requires it still
requires it to exist. It will just be set to `'standard'` if the
implementation does not support `'extended'`.

The only other case is when the canvas has not been configured
in which case `getConfiguration` returns undefined.

Both `getConfiguration` and `toneMapping` were added to the spec
close enough that any implemetation that supports `getConfiguration`
should be requried to include `toneMapping`.