Fix the iframe background color
I think this works. Originally when I did this 8 months ago
I set the CSS a certain way that works in Chrome but not Firefox.
I filed a bug for Firefox

https://bugzilla.mozilla.org/show_bug.cgi?id=1885048

They pointed out the bug is in Chrome and Firefox is correct.
I put the CSS to match what Firefox said was correct and found
the corresponding bug in Chrome which was basically being ignored.
(I can't seem to find the bug now)

Anyway, my understanding is, since the web started, iframes content
gets the background color of the iframe element. Unless of course
they set their own background color and that, with white being the
default backgrounc color, most pages expect a white background and
so iframe elements need have a white background otherwise content
breaks.

Often colors are now set via `color-scheme` but color-scheme is
ignored in the iframe content if its the same as the iframe element's
color-scheme because otherwise the previous rule about background color
would be un-enforced.

The solution appears to be, set the iframe element's color-scheme
to "initial". Now the iframe's content color-scheme setting is not
ignored.