Add a stencil mask example
I'm not sure this is a great sample because I'm not sure using
the stencil mask for this purpose is best. It's got the advantage
that it should "early out" but it has the disadvantage of, for example,
using an alpha mask, that would allow more than just masked yes/no.

Stencil masks are used for shadow volumes but that techinque has
fallen out of favor. Stencil masks are also used for outlines.
That might be a good use but there are several other popular
outline algorithmns.

So, this is demo is kind of pretty and at least shows using the
a stencil buffer.