Add Primitive Picking sample (#539)
Writes the `primitive_index` to a secondary render target and uses that
to highlight the primitive that the pointer is over on subsequent
frames. Also includes a mode that visualizes the primitive index. Used
as a (somewhat contrived) example of using the `primitive_index`
builtin.

This was based on the deferred shading sample, but has been modified
significantly.

---------

Co-authored-by: François Beaufort <beaufort.francois@gmail.com>